fix: Create a writer managing surrogate unicode
Unicode may have high value that overflows the 16bits range, java
encodes char using 16bits, thus those high unicode are spitted in two
16bits chars called surrogate form. Unfortunately, aapt2 do not handle
unicode in surrogate form and crash when apktool try to repack the res.
This writer perform the unicode logic that is added in ToStream in the
recent version of xalan, which is currently not added to any sdk of
android.